==> 1_preprocessing_real.ipynb <==
import pandas as pd
import numpy as np

loan_result = pd.read_csv("./data/loan_result.csv")
log_data = pd.read_csv("./data/log_data.csv")
user_spec = pd.read_csv("./data/user_spec.csv")

check_applied_ratio = loan_result[loan_result["loanapply_insert_time"] < "2022-06-01"]

check_applied_ratio["is_applied"].value_counts()

9709675/(9709675 + 560449)

560449/(9709675 + 560449)

kk = loan_result[loan_result["loanapply_insert_time"] >= "2022-06-01"]

kk[kk["application_id"] == 1882657]

#test = loan_result[loan_result["loanapply_insert_time"] >= "2022-06-01 00:00:00"]
#loan_result = loan_result[loan_result["loanapply_insert_time"] < "2022-06-01 00:00:00"]


### Step 1. Merge Dataset 

- 1. Drop columns : [personal_rehabilitation_yn , personal_rehabilitation_complete_yn ]
    - 위 컬럼들은 애초에 결측치가 너무 많고 추정하기가 쉽지않음.

loan_result.merge(user_spec).isna().sum()

#test = test.merge(user_spec)
loan_result = loan_result.merge(user_spec)

#test = test.drop(["personal_rehabilitation_yn" , "personal_rehabilitation_complete_yn" ], axis = 1)
loan_result = loan_result.drop(["personal_rehabilitation_yn" , "personal_rehabilitation_complete_yn" ], axis = 1)


### Step 2. Imputation

- loan_limit이 nan이면 loan_rate 역시 nan임
- 이 결측치들은 각 application_id의 평균값으로 대체

unique_loan = loan_result[loan_result['loan_limit'].isnull()]["application_id"].unique()

for i in range(len(unique_loan)):
    mean = loan_result[loan_result["application_id"] == unique_loan[i]].loc[:,"loan_limit"].dropna().mean()
    loan_result.loc[loan_result[loan_result["application_id"] == unique_loan[i]].loc[:,"loan_limit"].fillna( mean).index, "loan_limit"] = loan_result[loan_result["application_id"] == unique_loan[i]].loc[:,"loan_limit"].fillna( mean)

unique_loan = loan_result[loan_result['loan_rate'].isnull()]["application_id"].unique()

for i in range(len(unique_loan)):
    mean = loan_result[loan_result["application_id"] == unique_loan[i]].loc[:,"loan_rate"].dropna().mean()
    loan_result.loc[loan_result[loan_result["application_id"] == unique_loan[i]].loc[:,"loan_rate"].fillna( mean).index, "loan_rate"] = loan_result[loan_result["application_id"] == unique_loan[i]].loc[:,"loan_rate"].fillna( mean)

loan_result.isna().sum()

### 2-1 결측치 추정

#### 1.1 Birty_year & Gender 결측치 추정
- 기존에 있으면 이걸로 대체 (이건 전부 결측치임)

birth_data = loan_result[loan_result['birth_year'].isnull()]["application_id"].unique()

for i in range(len(birth_data)):
    mean = loan_result[loan_result["application_id"] == birth_data[i]].loc[:,"birth_year"].dropna().mean()
    loan_result.loc[loan_result[loan_result["application_id"] == birth_data[i]].loc[:,"birth_year"].fillna( mean).index, "birth_year"] = loan_result[loan_result["application_id"] == birth_data[i]].loc[:,"birth_year"].fillna(mean)

gender_data = loan_result[loan_result['gender'].isnull()]["application_id"].unique()

for i in range(len(gender_data)):
    mean = loan_result[loan_result["application_id"] == gender_data[i]].loc[:,"gender"].dropna().mean()
    loan_result.loc[loan_result[loan_result["application_id"] == gender_data[i]].loc[:,"gender"].fillna( mean).index, "gender"] = loan_result[loan_result["application_id"] == gender_data[i]].loc[:,"gender"].fillna(mean)

loan_result.isna().sum()

### 2-1 결측치 추정

#### 1.2 Credit_score 결측치 추정
- 기존에 있으면 이걸로 대체 
- 2022.06월 이후는 다 결측치인가본데..?

credit_data = loan_result[loan_result['credit_score'].isnull()]["application_id"].unique()

for i in range(len(credit_data)):
    mean = loan_result[loan_result["application_id"] == credit_data[i]].loc[:,"credit_score"].dropna().mean()
    loan_result.loc[loan_result[loan_result["application_id"] == credit_data[i]].loc[:,"credit_score"].fillna( mean).index, "credit_score"] = loan_result[loan_result["application_id"] == credit_data[i]].loc[:,"credit_score"].fillna(mean)

loan_result.isna().sum()

### 1.6 company_enter_month

- 근속년수를 파악할 수 있는 중요한 도구니깐.. 컬럼을 제거하면 안됨.
- 그런데 이걸 추정할 수 있는 마땅한 방법이 없음... 대체적으로 프리랜서인경우가 이거에 속할텐데 어떻게 해야할지 고민해야함
- company_enter_month 값들은 모두 0으로 바꾸고, 근속년수를 계산하기

loan_result_2 = pd.read_csv("./data/loan_result.csv")

loan_result_2 = loan_result_2.merge(user_spec)

loan_result["company_enter_month"] = loan_result_2["company_enter_month"]

# 최소 17살 이상일해야함
loan_result.loc[loan_result[((loan_result["birth_year"] * 100)+ 1700)>= loan_result["company_enter_month"] ].index, "company_enter_month"] = np.nan

loan_result.loc[loan_result[loan_result['company_enter_month'].isnull()].index, "company_enter_month"] = "000000"

loan_result['company_enter_month'] = loan_result['company_enter_month'].astype(float)

loan_result['company_enter_month'] = loan_result['company_enter_month'].astype(str).str[:6]

loan_result['company_enter_month'] = loan_result['company_enter_month'].astype(float).astype(int)

loan_result.loc[loan_result[loan_result['company_enter_month'] == 202211].index, "company_enter_month"] = 0

loan_result['company_enter_month'] = (202206 - loan_result['company_enter_month'])

348 / 12

def company_enter_enc(x):
    if x <= 3: return 0
    elif x <= 6: return 1
    elif x <= 9: return 2
    elif x <= 12: return 3
    elif x <= 24: return 4
    elif x <= 36: return 5
    elif x <= 48: return 6
    elif x <= 60: return 7
    elif x <= 72: return 8
    elif x <= 84: return 9
    elif x <= 96: return 10
    elif x <= 108: return 11
    elif x <= 120: return 12
    elif x <= 132: return 13
    elif x <= 144: return 14
    elif x <= 168: return 15
    elif x <= 180: return 16
    elif x <= 192: return 17
    elif x <= 204: return 18
    elif x <= 216: return 19
    elif x <= 228: return 20
    elif x <= 240: return 21
    elif x <= 252: return 22
    elif x <= 264: return 23
    elif x <= 276: return 24
    elif x <= 288: return 25
    elif x <= 300: return 26
    elif x <= 312: return 27
    elif x <= 324: return 28
    elif x <= 336: return 29
    elif x <= 348: return 30
    elif x <= 360: return 31

    else: return 99


loan_result["company_enter_month"] = loan_result["company_enter_month"].map(company_enter_enc)

### 1.8. existing_loan_cnt, existing_loan_amt

loan_result.loc[loan_result['existing_loan_cnt'] != loan_result['existing_loan_cnt'], 'existing_loan_cnt'] = 0

loan_result.loc[loan_result['existing_loan_amt'] != loan_result['existing_loan_amt'], 'existing_loan_amt'] = 0

loan_result.isna().sum()

### 1.9 Credit Score 다시생각해보기 - 일단제거

loan_result.loc[loan_result['purpose'] == "SWITCHLOAN", 'purpose'] = "대환대출"
loan_result.loc[loan_result['purpose'] == "LIVING", 'purpose'] = "생활비"
loan_result.loc[loan_result['purpose'] == "ETC", 'purpose'] = "기타"
loan_result.loc[loan_result['purpose'] == "BUYCAR", 'purpose'] = "자동차구입"
loan_result.loc[loan_result['purpose'] == "BUSINESS", 'purpose'] = "사업자금"
loan_result.loc[loan_result['purpose'] == "HOUSEDEPOSIT", 'purpose'] = "전월세보증금"
loan_result.loc[loan_result['purpose'] == "INVEST", 'purpose'] = "투자"
loan_result.loc[loan_result['purpose'] == "BUYHOUSE", 'purpose'] = "주택구입"


loan_result["income_type"].unique()

### Data Merge

test = loan_result[loan_result["loanapply_insert_time"] >= "2022-06-01 00:00:00"]
loan_result = loan_result[loan_result["loanapply_insert_time"] < "2022-06-01 00:00:00"]


test.to_csv("./prepro_data/submit_test.csv", index = False)

loan_result.to_csv("./prepro_data/full_data.csv", index = False)

test_set = pd.read_csv("./prepro_data/submit_test.csv")

test_set

3257239 

### 고민해볼것들

k = loan_result.groupby("product_id").mean()["is_applied"] * 100
k = pd.DataFrame(k).reset_index().sort_values("is_applied")

kk = loan_result.groupby("product_id").median()["is_applied"] * 100
kk = pd.DataFrame(kk).reset_index().sort_values("is_applied")

k[k["is_applied"] >= 10]

k = loan_result.groupby("bank_id").mean()["is_applied"] * 100
k = pd.DataFrame(k).reset_index().sort_values("is_applied")

k[k["is_applied"] >= 10]

loan_result[loan_result["product_id"] == 244]

k = loan_result.groupby("existing_loan_cnt").mean()["is_applied"]

k = pd.DataFrame(k).reset_index().sort_values("is_applied")

k[k["existing_loan_cnt"] > 4]["is_applied"].mean()

loan_result["existing_loan_cnt"].describe()

loan_result.columns

check_the_set_cover = loan_result[["application_id", "loanapply_insert_time", 'loan_limit', 'loan_rate', 'is_applied']]

check_the_set_cover[check_the_set_cover["application_id"] == check_the_set_cover["application_id"].unique()[13]]
