==> 5_test_modeling-ACC-ALL.ipynb <==
from glob import glob
from tqdm import tqdm
#from rdkit import Chem
import numpy as np
import pandas as pd
import re
import shutil
import os

import autogluon
from autogluon.common.utils.utils import setup_outputdir
from autogluon.tabular import TabularDataset, TabularPredictor
from autogluon.core.utils.loaders import load_pkl
from autogluon.core.utils.savers import save_pkl
import os.path
import sklearn
from sklearn.model_selection import train_test_split
import autogluon.core as ag
import pandas as pd
import numpy as np

import pandas as pd
import random
import argparse
import os
from sklearn.preprocessing import LabelEncoder
from sklearn.pipeline import Pipeline
import numpy as np
from sklearn.model_selection import StratifiedShuffleSplit

import glob

import datetime 
import time     

full_data = pd.read_csv("./dataset/full_data.csv")

full_data.info()

# le_income_type

le_purpose_type = LabelEncoder()
le_purpose_type = le_purpose_type.fit(full_data["purpose"])
le_purpose_fitted = le_purpose_type.fit_transform(full_data["purpose"])
purpose_inverted = le_purpose_type.inverse_transform(le_purpose_fitted)

print(le_purpose_type.classes_)

full_data["purpose"] = le_purpose_fitted

full_data["bank_id"] = full_data["bank_id"].astype('category')
full_data["product_id"] = full_data["product_id"].astype('category')
full_data["loan_limit"] = full_data["loan_limit"].astype('category')
full_data["birth_year"] = full_data["birth_year"].astype('category')

full_data["gender"] = full_data["gender"].astype('category')
full_data["credit_score"] = full_data["credit_score"].astype('category')
full_data["yearly_income"] = full_data["yearly_income"].astype('category')
full_data["income_type"] = full_data["income_type"].astype('category')
full_data["company_enter_month"] = full_data["company_enter_month"].astype('category')

full_data["employment_type"] = full_data["employment_type"].astype('category')
full_data["houseown_type"] = full_data["houseown_type"].astype('category')
full_data["desired_amount"] = full_data["desired_amount"].astype('category')
full_data["existing_loan_cnt"] = full_data["existing_loan_cnt"].astype('category')
full_data["existing_loan_amt"] = full_data["existing_loan_amt"].astype('category')

full_data["credit_score_nan"] = full_data["credit_score_nan"].astype('category')
full_data["gender_birth_nan"] = full_data["gender_birth_nan"].astype('category')
full_data["loan_limit/max_loan_limit_binary"] = full_data["loan_limit/max_loan_limit_binary"].astype('category')
full_data["loan_rate/min_loan_rate_binary"] = full_data["loan_rate/min_loan_rate_binary"].astype('category')

full_data["desired_amount/existing_loan_amt_binary"] = full_data["desired_amount/existing_loan_amt_binary"].astype('category')
full_data["hour"] = full_data["hour"].astype('category')
full_data["dayofweek"] = full_data["dayofweek"].astype('category')
full_data["purpose"] = full_data["purpose"].astype('category')
full_data["existing_loan_cnt_binary"] = full_data["existing_loan_cnt_binary"].astype("category")

full_data.info()

full_data["purpose"].unique()

full_data = full_data.drop(["application_id", "user_id"], axis = 1)

### K fold ( Set 5 Fold)

def k_fold_result(data):
    ## Test set setting
    data = data.reset_index().iloc[:, 1:]
    
    train, test = train_test_split(data, test_size=0.15,   random_state=20205289,  stratify=data['is_applied'])

    train = train.reset_index().iloc[:, 1:]
    test = test.reset_index().iloc[:, 1:]

    from sklearn.model_selection import StratifiedKFold
    kf = StratifiedKFold(n_splits=5, shuffle=True, random_state=20205289)
    for fold, (train_idx, val_idx) in enumerate(kf.split(range(len(train)), y=train['is_applied'])):
        train.loc[val_idx, 'fold'] = fold
        
    return train, test

### Modeling

def modeling(dataset, fold):
    
    dataset, test = k_fold_result(dataset)
    test.to_csv(f"./fold_data/model_all_test.csv")
    
    train = dataset.reset_index().iloc[:, 1:]
    train["is_applied"] = train["is_applied"].astype(int)
    
    for i in range(0, fold):

        val_idx = list(train[train['fold'] == int(i)].index)
        val_df = train[train['fold'] == i].reset_index(drop=True).iloc[:, :-1]
        train_df = train[train['fold'] != i].reset_index(drop=True).iloc[:, :-1]

        labels = 'is_applied'
        problem_types = 'binary'  
        save_path = f'model_test_all/model_all_fold_{i}'  

        excluded_model_types = ['KNN', "NN_TORCH", "FASTAI", "CAT", "Cat"]
        multi_predictor = TabularPredictor(label=labels, problem_type=problem_types, eval_metric="f1", path=save_path)
        multi_predictor.fit(train_df, tuning_data=val_df, excluded_model_types=excluded_model_types, hyperparameters = "light") #['default', 'light', 'very_light', 'toy', 'multimodal', 'interpretable']
        
        train_df.to_csv(f"./fold_data/model_all_fold_{i}.csv")
        val_df.to_csv(f"./fold_data/model_all_fold_{i}.csv")
        
        time.sleep(60)
        print("TIME")
    
    return test
    

#'생활비', '대환대출', '사업자금', '투자', '기타', '전월세보증금', '주택구입', '자동차구입'
test_df = modeling(dataset = full_data, fold = 5)

dataset, test = k_fold_result(full_data)

train = dataset.reset_index().iloc[:, 1:]
train["is_applied"] = train["is_applied"].astype(int)


val_idx_0 = list(train[train['fold'] == 0].index)
val_df_0 = train[train['fold'] == 0].reset_index(drop=True).iloc[:, :-1]
train_df_0 = train[train['fold'] != 0].reset_index(drop=True).iloc[:, :-1]

val_idx_1 = list(train[train['fold'] == 1].index)
val_df_1 = train[train['fold'] == 1].reset_index(drop=True).iloc[:, :-1]
train_df_1 = train[train['fold'] != 1].reset_index(drop=True).iloc[:, :-1]

val_idx_2 = list(train[train['fold'] == 2].index)
val_df_2 = train[train['fold'] == 2].reset_index(drop=True).iloc[:, :-1]
train_df_2 = train[train['fold'] != 2].reset_index(drop=True).iloc[:, :-1]

val_idx_3 = list(train[train['fold'] == 3].index)
val_df_3 = train[train['fold'] == 3].reset_index(drop=True).iloc[:, :-1]
train_df_3 = train[train['fold'] != 3].reset_index(drop=True).iloc[:, :-1]

val_idx_4 = list(train[train['fold'] == 4].index)
val_df_4 = train[train['fold'] == 4].reset_index(drop=True).iloc[:, :-1]
train_df_4 = train[train['fold'] != 4].reset_index(drop=True).iloc[:, :-1]


### Test

reg_01 = TabularPredictor.load("model_test_all/model_all_fold_0")
reg_02 = TabularPredictor.load("model_test_all/model_all_fold_1")
reg_03 = TabularPredictor.load("model_test_all/model_all_fold_2")
reg_04 = TabularPredictor.load("model_test_all/model_all_fold_3")
reg_05 = TabularPredictor.load("model_test_all/model_all_fold_4")


name_01 = reg_01.leaderboard(val_df_0, silent=True)
name_02 = reg_02.leaderboard(val_df_1, silent=True)
name_03 = reg_03.leaderboard(val_df_2, silent=True)
name_04 = reg_04.leaderboard(val_df_3, silent=True)
name_05 = reg_05.leaderboard(val_df_4, silent=True)


print("fold_0 : ", name_01["model"][0])
print("fold_1 : ", name_02["model"][0])
print("fold_2 : ", name_03["model"][0])
print("fold_3 : ", name_04["model"][0])
print("fold_4 : ", name_05["model"][0])

rag_01_model_name = reg_01._trainer.load_model(name_01["model"][0])
rag_02_model_name = reg_02._trainer.load_model(name_02["model"][0])
rag_03_model_name = reg_03._trainer.load_model(name_03["model"][0])
rag_04_model_name = reg_04._trainer.load_model(name_04["model"][0])
rag_05_model_name = reg_05._trainer.load_model(name_05["model"][0])

dataset, test_df = k_fold_result(full_data)

### Evaluation

rag_01_model_p  = reg_01.predict_proba(test_df, model = rag_01_model_name )
rag_02_model_p  = reg_02.predict_proba(test_df, model = rag_02_model_name )
rag_03_model_p  = reg_03.predict_proba(test_df, model = rag_03_model_name )
rag_04_model_p  = reg_04.predict_proba(test_df, model = rag_04_model_name )
rag_05_model_p  = reg_05.predict_proba(test_df, model = rag_05_model_name )
