==> 6_DL_models_inputs.ipynb <==
# Application을 기준으로 대출했는지 안했는지만 체크

from glob import glob
from tqdm import tqdm
#from rdkit import Chem
import numpy as np
import pandas as pd
import re
import shutil
import os

import autogluon
from autogluon.common.utils.utils import setup_outputdir
from autogluon.tabular import TabularDataset, TabularPredictor
from autogluon.core.utils.loaders import load_pkl
from autogluon.core.utils.savers import save_pkl
import os.path
import sklearn
from sklearn.model_selection import train_test_split
import autogluon.core as ag
import pandas as pd
import numpy as np

import pandas as pd
import random
import argparse
import os
from sklearn.preprocessing import LabelEncoder
from sklearn.pipeline import Pipeline
import numpy as np
from sklearn.model_selection import StratifiedShuffleSplit

import glob

full_data = pd.read_csv("./dataset/full_data.csv")

# le_income_type

le_purpose_type = LabelEncoder()
le_purpose_type = le_purpose_type.fit(full_data["purpose"])
le_purpose_fitted = le_purpose_type.fit_transform(full_data["purpose"])
purpose_inverted = le_purpose_type.inverse_transform(le_purpose_fitted)

print(le_purpose_type.classes_)

full_data["purpose"] = le_purpose_fitted

full_data["bank_id"] = full_data["bank_id"].astype('category')
full_data["product_id"] = full_data["product_id"].astype('category')
full_data["loan_limit"] = full_data["loan_limit"].astype('category')
full_data["birth_year"] = full_data["birth_year"].astype('category')

full_data["gender"] = full_data["gender"].astype('category')
full_data["credit_score"] = full_data["credit_score"].astype('category')
full_data["yearly_income"] = full_data["yearly_income"].astype('category')
full_data["income_type"] = full_data["income_type"].astype('category')
full_data["company_enter_month"] = full_data["company_enter_month"].astype('category')

full_data["employment_type"] = full_data["employment_type"].astype('category')
full_data["houseown_type"] = full_data["houseown_type"].astype('category')
full_data["desired_amount"] = full_data["desired_amount"].astype('category')
full_data["existing_loan_cnt"] = full_data["existing_loan_cnt"].astype('category')
full_data["existing_loan_amt"] = full_data["existing_loan_amt"].astype('category')

full_data["credit_score_nan"] = full_data["credit_score_nan"].astype('category')
full_data["gender_birth_nan"] = full_data["gender_birth_nan"].astype('category')
full_data["loan_limit/max_loan_limit_binary"] = full_data["loan_limit/max_loan_limit_binary"].astype('category')
full_data["loan_rate/min_loan_rate_binary"] = full_data["loan_rate/min_loan_rate_binary"].astype('category')

full_data["desired_amount/existing_loan_amt_binary"] = full_data["desired_amount/existing_loan_amt_binary"].astype('category')
full_data["hour"] = full_data["hour"].astype('category')
full_data["dayofweek"] = full_data["dayofweek"].astype('category')
full_data["purpose"] = full_data["purpose"].astype('category')
full_data["existing_loan_cnt_binary"] = full_data["existing_loan_cnt_binary"].astype("category")

full_data = full_data.drop(["application_id", "user_id"], axis = 1)

full_data.info()

'loan_rate', 'loan_limit/max_loan_limit', 'loan_rate/min_loan_rate', 'desired_amount/existing_loan_amt'

full_data.loc[full_data[full_data['desired_amount/existing_loan_amt'] >= 3].index, 'desired_amount/existing_loan_amt'] = 3

print("loan_rate")
print("q2",full_data['loan_rate'].median())
print("q3", full_data['loan_rate'].quantile(0.75))
print("q1", full_data['loan_rate'].quantile(0.25))
print("==")
print("loan_limit/max_loan_limit")
print("q2",full_data['loan_limit/max_loan_limit'].median())
print("q3", full_data['loan_limit/max_loan_limit'].quantile(0.75))
print("q1", full_data['loan_limit/max_loan_limit'].quantile(0.25))
print("==")
print("loan_rate/min_loan_rate")
print("q2",full_data['loan_rate/min_loan_rate'].median())
print("q3", full_data['loan_rate/min_loan_rate'].quantile(0.75))
print("q1", full_data['loan_rate/min_loan_rate'].quantile(0.25))
print("==")
print("desired_amount/existing_loan_amt")
print("q2",full_data['desired_amount/existing_loan_amt'].median())
print("q3", full_data['desired_amount/existing_loan_amt'].quantile(0.75))
print("q1", full_data['desired_amount/existing_loan_amt'].quantile(0.25))

full_data['loan_rate'] = (full_data['loan_rate'] - full_data['loan_rate'].median() )/ (full_data['loan_rate'].quantile(0.75) - full_data['loan_rate'].quantile(0.25))
full_data['loan_limit/max_loan_limit'] = (full_data['loan_limit/max_loan_limit'] - full_data['loan_limit/max_loan_limit'].median() )/ (full_data['loan_limit/max_loan_limit'].quantile(0.75) - full_data['loan_limit/max_loan_limit'].quantile(0.25))
full_data['loan_rate/min_loan_rate'] = (full_data['loan_rate/min_loan_rate'] - full_data['loan_rate/min_loan_rate'].median() )/ (full_data['loan_rate/min_loan_rate'].quantile(0.75) - full_data['loan_rate/min_loan_rate'].quantile(0.25))

print("loan_rate")
print("q2",full_data['loan_rate'].median())
print("q3", full_data['loan_rate'].quantile(0.75))
print("q1", full_data['loan_rate'].quantile(0.25))
print("==")
print("loan_limit/max_loan_limit")
print("q2",full_data['loan_limit/max_loan_limit'].median())
print("q3", full_data['loan_limit/max_loan_limit'].quantile(0.75))
print("q1", full_data['loan_limit/max_loan_limit'].quantile(0.25))
print("==")
print("loan_rate/min_loan_rate")
print("q2",full_data['loan_rate/min_loan_rate'].median())
print("q3", full_data['loan_rate/min_loan_rate'].quantile(0.75))
print("q1", full_data['loan_rate/min_loan_rate'].quantile(0.25))
print("==")
print("desired_amount/existing_loan_amt")
print("q2",full_data['desired_amount/existing_loan_amt'].median())
print("q3", full_data['desired_amount/existing_loan_amt'].quantile(0.75))
print("q1", full_data['desired_amount/existing_loan_amt'].quantile(0.25))

full_data.columns

print("bank_id :" , len(full_data["bank_id"].unique()) + 1)
print("product_id : " , len(full_data["product_id"].unique()) + 1)
print("loan_limit : " , len(full_data["loan_limit"].unique()) + 1)
print("birth_year : " , len(full_data["birth_year"].unique()) + 1)
print("gender : " , len(full_data["gender"].unique()) + 1)
print("credit_score : " , len(full_data["credit_score"].unique()) + 1)
print("yearly_income : " , len(full_data["yearly_income"].unique()) + 1)
print("income_type : " , len(full_data["income_type"].unique()) + 1)
print("company_enter_month : " , len(full_data["company_enter_month"].unique()) + 1)
print("employment_type : " , len(full_data["employment_type"].unique()) + 1)
print("houseown_type : " , len(full_data["houseown_type"].unique()) + 1)
print("desired_amount : " , len(full_data["desired_amount"].unique()) + 1)
print("purpose : " , len(full_data["purpose"].unique()) + 1)
print("existing_loan_cnt : " , len(full_data["existing_loan_cnt"].unique()) + 1)
print("existing_loan_amt : " , len(full_data["existing_loan_amt"].unique()) + 1)
print("credit_score_nan : " , len(full_data["credit_score_nan"].unique()) + 1)
print("gender_birth_nan : " , len(full_data["gender_birth_nan"].unique()) + 1)
print("existing_loan_cnt_binary : " , len(full_data["existing_loan_cnt_binary"].unique()) + 1)
print("loan_limit/max_loan_limit_binary : " , len(full_data["loan_limit/max_loan_limit_binary"].unique()) + 1)
print("loan_rate/min_loan_rate_binary : " , len(full_data["loan_rate/min_loan_rate_binary"].unique()) + 1)
print("desired_amount/existing_loan_amt_binary : " , len(full_data["desired_amount/existing_loan_amt_binary"].unique()) + 1)
print("hour : " , len(full_data["hour"].unique()) + 1)
print("dayofweek : " , len(full_data["dayofweek"].unique()) + 1)
#print("existing_loan_cnt_binary: ", len(full_data["existing_loan_cnt_binary"].unique()) + 1)

full_data_cat = full_data[["bank_id", "product_id", "loan_limit", "birth_year", "gender", "credit_score", "yearly_income", "income_type", "company_enter_month",\
                  "employment_type", "houseown_type", "desired_amount", "purpose", "existing_loan_cnt", "existing_loan_amt" , "credit_score_nan", "gender_birth_nan",\
                  "existing_loan_cnt_binary", "loan_limit/max_loan_limit_binary", "loan_rate/min_loan_rate_binary", "desired_amount/existing_loan_amt_binary", "hour", "dayofweek"]]

full_data_cont = full_data[['loan_rate', 'loan_limit/max_loan_limit', 'loan_rate/min_loan_rate', 'desired_amount/existing_loan_amt', 'num_not_approve', 'num_total_approve', 'num_approve', 'num_total_login']]

full_data_target = full_data[["is_applied"]]

full_data = pd.concat([full_data_cont, full_data_cat, full_data_target], axis = 1)

full_data

def k_fold_result(data):
    ## Test set setting
    data = data.reset_index().iloc[:, 1:]
    

    #train, test = train_test_split(data, test_size=0.15,   random_state=20205289,  stratify=data['is_applied'])

    train = data.reset_index().iloc[:, 1:]
    #test = test.reset_index().iloc[:, 1:]

    from sklearn.model_selection import StratifiedKFold
    kf = StratifiedKFold(n_splits=5, shuffle=True, random_state=20205289)
    for fold, (train_idx, val_idx) in enumerate(kf.split(range(len(train)), y=train['is_applied'])):
        train.loc[val_idx, 'fold'] = fold
        
    return train

train = k_fold_result(full_data)

fold_0 = train[train["fold"] == 0].iloc[:, :-1].reset_index().iloc[:, 1:]
fold_1 = train[train["fold"] == 1].iloc[:, :-1].reset_index().iloc[:, 1:]
fold_2 = train[train["fold"] == 2].iloc[:, :-1].reset_index().iloc[:, 1:]
fold_3 = train[train["fold"] == 3].iloc[:, :-1].reset_index().iloc[:, 1:]
fold_4 = train[train["fold"] == 4].iloc[:, :-1].reset_index().iloc[:, 1:]

fold_0.to_csv("./DL_dataset/fold_0.csv", index = False)
fold_1.to_csv("./DL_dataset/fold_1.csv", index = False)
fold_2.to_csv("./DL_dataset/fold_2.csv", index = False)
fold_3.to_csv("./DL_dataset/fold_3.csv", index = False)
fold_4.to_csv("./DL_dataset/fold_4.csv", index = False)

train.to_csv("./DL_dataset/train.csv", index = False)

train["fold"].unique()

train.iloc[:, :8]

train.iloc[:, 8:-2]

train.shape

l=[]
for i in range(8, 31):
    print(train.iloc[:, i:i+1].columns)
    l.append(len(train.iloc[:, i].astype(str).unique()) + 1)

l

l == [63, 179, 52, 15, 4, 25, 20, 7, 10, 5, 5, 20, 9, 11, 20, 3, 3, 3, 3, 3, 4, 25, 8]

train.columns

### Submit dataset

submit_test = pd.read_csv("./dataset/submit_test.csv")

# le_income_type

le_purpose_type = LabelEncoder()
le_purpose_type = le_purpose_type.fit(full_data["purpose"])
le_purpose_fitted = le_purpose_type.fit_transform(submit_test["purpose"])
purpose_inverted = le_purpose_type.inverse_transform(le_purpose_fitted)

print(le_purpose_type.classes_)

submit_test["purpose"] = le_purpose_fitted

submit_test["bank_id"] = submit_test["bank_id"].astype('category')
submit_test["product_id"] = submit_test["product_id"].astype('category')
submit_test["loan_limit"] = submit_test["loan_limit"].astype('category')
submit_test["birth_year"] = submit_test["birth_year"].astype('category')
submit_test["gender"] = submit_test["gender"].astype('category')
submit_test["credit_score"] = submit_test["credit_score"].astype('category')
submit_test["yearly_income"] = submit_test["yearly_income"].astype('category')
submit_test["income_type"] = submit_test["income_type"].astype('category')
submit_test["company_enter_month"] = submit_test["company_enter_month"].astype('category')
submit_test["employment_type"] = submit_test["employment_type"].astype('category')
submit_test["houseown_type"] = submit_test["houseown_type"].astype('category')
submit_test["desired_amount"] = submit_test["desired_amount"].astype('category')
submit_test["existing_loan_cnt"] = submit_test["existing_loan_cnt"].astype('category')
submit_test["existing_loan_amt"] = submit_test["existing_loan_amt"].astype('category')
submit_test["credit_score_nan"] = submit_test["credit_score_nan"].astype('category')
submit_test["gender_birth_nan"] = submit_test["gender_birth_nan"].astype('category')
submit_test["loan_limit/max_loan_limit_binary"] = submit_test["loan_limit/max_loan_limit_binary"].astype('category')
submit_test["loan_rate/min_loan_rate_binary"] = submit_test["loan_rate/min_loan_rate_binary"].astype('category')
submit_test["desired_amount/existing_loan_amt_binary"] = submit_test["desired_amount/existing_loan_amt_binary"].astype('category')
submit_test["hour"] = submit_test["hour"].astype('category')
submit_test["dayofweek"] = submit_test["dayofweek"].astype('category')
submit_test["purpose"] = submit_test["purpose"].astype('category')
submit_test["existing_loan_cnt_binary"] = submit_test["existing_loan_cnt_binary"].astype("category")

submit_test = submit_test.drop(["application_id", "user_id"], axis = 1)

submit_test.loc[submit_test[submit_test['desired_amount/existing_loan_amt'] >= 3].index, 'desired_amount/existing_loan_amt'] = 3

submit_test['loan_rate'] = (submit_test['loan_rate'] - full_data['loan_rate'].median() )/ (full_data['loan_rate'].quantile(0.75) - full_data['loan_rate'].quantile(0.25))
submit_test['loan_limit/max_loan_limit'] = (submit_test['loan_limit/max_loan_limit'] - full_data['loan_limit/max_loan_limit'].median() )/ (full_data['loan_limit/max_loan_limit'].quantile(0.75) - full_data['loan_limit/max_loan_limit'].quantile(0.25))
submit_test['loan_rate/min_loan_rate'] = (submit_test['loan_rate/min_loan_rate'] - full_data['loan_rate/min_loan_rate'].median() )/ (full_data['loan_rate/min_loan_rate'].quantile(0.75) - full_data['loan_rate/min_loan_rate'].quantile(0.25))

submit_test_cat = submit_test[["bank_id", "product_id", "loan_limit", "birth_year", "gender", "credit_score", "yearly_income", "income_type", "company_enter_month",\
                  "employment_type", "houseown_type", "desired_amount", "purpose", "existing_loan_cnt", "existing_loan_amt" , "credit_score_nan", "gender_birth_nan",\
                  "existing_loan_cnt_binary", "loan_limit/max_loan_limit_binary", "loan_rate/min_loan_rate_binary", "desired_amount/existing_loan_amt_binary", "hour", "dayofweek"]]

submit_test_cont = submit_test[['loan_rate', 'loan_limit/max_loan_limit', 'loan_rate/min_loan_rate', 'desired_amount/existing_loan_amt', 'num_not_approve', 'num_total_approve', 'num_approve', 'num_total_login']]

submit_test_cont_target = submit_test[["is_applied"]]

submit_test = pd.concat([submit_test_cont, submit_test_cat, submit_test_cont_target], axis = 1)

submit_test["is_applied"] = 0

submit_test.columns

train.columns

submit_test.to_csv("./DL_dataset/test.csv", index = False)

submit_test["purpose"]
