==> 9_select_the_submit_data.ipynb <==
import pandas as pd

submit_test = pd.read_csv("./submit/test.csv")
submit_data = pd.read_csv("submit_data.csv")

loan_result = pd.read_csv("./data/loan_result.csv")
user_spec = pd.read_csv("./data/user_spec.csv")

submit_test_origin = pd.read_csv("./data/loan_result.csv")
submit_test_origin = submit_test_origin[submit_test_origin["loanapply_insert_time"] >= "2022-06-01"].reset_index().iloc[:, 1:]

DL_Dataset = pd.read_csv("./DL_dataset/test.csv")

loan_result = loan_result[loan_result["loanapply_insert_time"] >= "2022-06-01"].reset_index().iloc[:, 1:]
loan_result = loan_result.reset_index()
loan_result = loan_result.merge(user_spec)

submit_test["index"] = loan_result["index"]
submit_test = submit_test.sort_values("index").reset_index().iloc[:, 1:]
submit_test = submit_test.drop("index", axis = 1)


not_selected = list(set(submit_test["application_id"].values).difference(set(submit_data["application_id"])))
submit_test = submit_test[~submit_test["application_id"].isin(not_selected)]

not_selected = list(set(submit_test["product_id"].values).difference(set(submit_data["product_id"])))
submit_test = submit_test[~submit_test["product_id"].isin(not_selected)]

submit_test["merge"] = submit_test["application_id"].astype(str) + "_" + submit_test["product_id"].astype(str)
submit_data["merge"] = submit_data["application_id"].astype(str) + "_" + submit_data["product_id"].astype(str)

not_selected = list(set(submit_test["merge"].values).difference(set(submit_data["merge"])))
submit_test = submit_test[~submit_test["merge"].isin(not_selected)]

submit_test = submit_test.drop_duplicates(["application_id", "product_id"], keep = "first")
submit_test = submit_test.drop("merge", axis = 1)

submit_data = submit_data.drop("merge", axis = 1)

submit_test = submit_data[["application_id", "product_id"]].merge(submit_test)

(submit_test["application_id"].values  == submit_data["application_id"].values).sum()

(submit_test["product_id"].values  == submit_data["product_id"].values).sum()

submit_test.reset_index().iloc[:, 1:].to_csv("./submit/submission.csv", index = False)
